Updated and fixed setup-* scripts per #23

+++ setup_raspbian64.sh
@@ -157,7 +157,6 @@ else
     fi
 fi
 
-exit 0
 
 if yesno "Do you have a hardware (RTC) clock module"; then
 	DO_RTC=Y
@@ -213,7 +212,7 @@ HW_CLOCK_SET="
 # Reset the System Clock to UTC if the hardware clock from which it
 # was copied by the kernel was in localtime.
 
-dev=$1
+dev=\$1
 
 #if [ -e /run/systemd/system ] ; then
 #    exit 0
@@ -236,12 +235,12 @@ if [ -f /etc/default/hwclock ] ; then
     . /etc/default/hwclock
 fi
 
-if [ yes = "$BADYEAR" ] ; then
-    /sbin/hwclock --rtc=$dev --systz --badyear
-    /sbin/hwclock --rtc=$dev --hctosys --badyear
+if [ yes = \"\$BADYEAR\" ] ; then
+    /sbin/hwclock --rtc=\$dev --systz --badyear
+    /sbin/hwclock --rtc=\$dev --hctosys --badyear
 else
-    /sbin/hwclock --rtc=$dev --systz
-    /sbin/hwclock --rtc=$dev --hctosys
+    /sbin/hwclock --rtc=\$dev --systz
+    /sbin/hwclock --rtc=\$dev --hctosys
 fi
 
 # Note 'touch' may not be available in initramfs
@@ -256,7 +255,7 @@ Description=Enable ${RTC_CHIP} I2C RTC
 
 [Service]
 Type=oneshot
-ExecStartPre=/bin/bash -c "echo ${RTC_CHIP} 0x68 | tee /sys/class/i2c-adapter/i2c-1/new_device"
+ExecStartPre=/bin/bash -c \"echo ${RTC_CHIP} 0x68 | tee /sys/class/i2c-adapter/i2c-1/new_device\"
 ExecStart=/sbin/hwclock -s
 
 [Install]
@@ -287,6 +286,9 @@ sudo apt-get install -y libnss-mdns avahi-utils libavahi-compat-libdnssd-dev
 sudo systemctl stop systemd-timesyncd
 sudo systemctl disable systemd-timesyncd
 sudo apt-get install -y ntp
+#force time sync
+sudo service ntp stop
+sudo ntpd -gq
 sudo service ntp start
 
 # routing
@@ -330,7 +332,9 @@ else
 	sudo sed -i 's/dtoverlay=i2c-rtc,pcf8523//' /boot/config.txt
 	sudo sed -i 's/dtoverlay=i2c-rtc,ds3231//' /boot/config.txt
 	sudo apt-get -y install fake-hwclock
-	sudo cp /lib/udev/hwclock-set.orig /lib/udev/hwclock-set
+	if [ -e /lib/udev/hwclock-set.orig ]; then
+		sudo cp /lib/udev/hwclock-set.orig /lib/udev/hwclock-set
+	fi
 fi
 
 curl -sL https://repos.influxdata.com/influxdb.key | sudo apt-key add -
@@ -340,6 +344,11 @@ sudo apt update
 
 sudo apt-get install -y influxdb
 sudo sed -i 's/store-enabled = true/store-enabled = false/' /etc/influxdb/influxdb.conf
+sudo sed -i 's/log-enabled = true/log-enabled = false/' /etc/influxdb/influxdb.conf
+sudo sed -i 's/# log-enabled = false/log-enabled = false/' /etc/influxdb/influxdb.conf
+	#max-connection-limit = 0
+	#max-enqueued-write-limit = 0
+	#enqueued-write-timeout = 3s
 sudo service influxdb restart
 
 if [ ! -f /tmp/bellsoft-jdk11.0.2-linux-aarch64-lite.tar.gz ]; then
